feat: 내정보 조회와 수정 레이어 분리
